requirements.txt
rich==13.7.1
requests==2.31.0
python-nmap==0.7.1
fpdf2==2.7.9
python-dotenv==1.0.1
